requirements.txt
pandas==1.3.5
opencv-python==4.6.0.66
yacs==0.1.8
tqdm
pickle5
Pillow==8.4.0
PyYAML==6.0
wandb==0.13.5
accelerate
ml-collections
matplotlib
timm
autopep8